### 2022-02-23 #4198 @editorialbot
[PRE REVIEW]: FunFact: Build Your Own Tensor Decomposition Model in a Breeze
**Submitting author:** @yhtang (<a href="http://orcid.org/0000-0001-7424-5439">Yu-Hang Tang</a>)
**Repository:** https://github.com/yhtang/FunFact
**Branch with paper.md** (empty if default branch): workspace/joss-submission
**Version:** 1.0.0
