+++ funfact/lang/interpreter/_evaluation.py
@@ -43,9 +43,10 @@ class Evaluator(ROOFInterpreter):
             shape, **kwargs):
         return _einop(lhs, rhs, einspec, shape)
 
-    def tran(self, src, indices, einspec, **kwargs):
-        in_spec, out_spec = einspec.split('->')
-        return ab.transpose(src, [in_spec.index(i) for i in out_spec])
+    def tran(self, src, indices, transpec, **kwargs):
+        return ab.transpose(
+            src, [transpec.in_idx.index(i) for i in transpec.out_idx]
+        )
 
     def abstract_dest(self, src, indices, **kwargs):
         raise NotImplementedError()
