+++ funfact/lang/interpreter/_latex.py
@@ -44,7 +44,7 @@ class LatexRenderer(ROOFInterpreter):
 
     def tensor(self, decl, **kwargs):
         return decl._repr_tex_()
-    
+
     def ellipsis(self, **kwargs):
         return '...'
 
