### 2022-02-28 f27794c
precompute order
