### 2022-02-28 b1166a8
remove ellipsis functionality from type deducer
