### 2022-03-07 #4198 @melissawm
Hi @yhtang

- it seems you are using a slightly modified version of the [BSD-3 clause LBNL License](https://opensource.org/BSD-3-Clause-LBNL) - can you remove the added text at the beginning of  the first line so that the standard license text is used?
- On line 99, the citation to Tensor Toolbox is malformed - it seems there is a "and " missing before "others";
- Can you clarify the citation for TensorOperations.jl? I see one author cited but several contributors in the github repository, and no citation instructions.
- Could you add the doi for TensorLy? https://dl.acm.org/doi/abs/10.5555/3322706.3322732

Thank you!
