### 2022-03-07 #4198 @yhtang
Hi @melissawm,

I tried to compare our license with the standard BSD one but could not figure out where the significant differences are. Could you please let us know which particular text in the license is causing the issue?
> the added text at the beginning of the first line

I'll then discuss this with our intellectual property office to see what the options are.

We'll take care of the rest of the action items shortly.

Thanks!
