### 2022-03-07 #4198 @yhtang
@melissawm Regarding
> Could you add the doi for TensorLy? https://dl.acm.org/doi/abs/10.5555/3322706.3322732

It appears unfortunately that there is no DOI associated with the article. Even though the URL reads like it does contain one, i.e. `10.5555/3322706.3322732`, looking it up on [doi.org](https://www.doi.org/) returns nothing. The paper's [JMLR page](https://www.jmlr.org/papers/v20/18-277.html) (and the `bib` link therein) also does not contain a DOI.
