### 2022-03-11 841e405
working on parametrized tensors
