+++ funfact/__init__.py
@@ -13,7 +13,7 @@ from .context import is_grad_on, enable_grad
 from . import initializers
 from . import conditions
 from .lang._semiring import minplus, logsumexp, viterbi
-
+from . import generators
 
 __all__ = [
     'use',
@@ -37,7 +37,8 @@ __all__ = [
     'conditions',
     'minplus',
     'logsumexp',
-    'viterbi'
+    'viterbi',
+    'generators'
 ]
 
 
