+++ funfact/lang/interpreter/_evaluation.py
@@ -18,6 +18,9 @@ class Evaluator(ROOFInterpreter):
     def literal(self, value, **kwargs):
         return ab.tensor(value.raw)
 
+    def parametrized_tensor(self, decl, data,  **kwargs):
+        return decl.generator(data)
+
     def tensor(self, decl, data, **kwargs):
         return data
 
