### 2022-03-11 0faad4f
fix generator vectorization
