+++ funfact/lang/interpreter/_elementwise.py
@@ -15,7 +15,10 @@ def _sliced_size(sli, sz):
 
 class ElementwiseEvaluator(Evaluator):
     def parametrized_tensor(self, decl, data, slices, **kwargs):
-        return decl.generator(data)[slices]
+        try:
+            return decl.generator(data, slices)
+        except TypeError:
+            return decl.generator(data)[slices]
 
     def tensor(self, decl, data, slices, **kwargs):
         return data[slices]
