### 2022-03-16 de9944f
improve rotation UI
