### 2022-03-17 #4198 @yhtang
> I believe it is just the mention of the project name that should be removed. This has come up other times with the LBNL license and one possible solution seems to be adding an extra Copyright file with that mention, if it is required. See for example [#2666 (comment)](https://github.com/openjournals/joss-reviews/issues/2666#issuecomment-691257600)

I just made the change after getting the go-ahead from our legal team. The latest of our submission branch `workspace/joss-submission` should now contain the conforming `LICENSE` file. :smile:
