### 2022-03-17 #4198 @yhtang
Still haven't heard back regarding https://github.com/Jutho/TensorOperations.jl/issues/117. On a side note, the only activity of the repo in the last month is this issue that we created.
