### 2022-03-17 #4198 @yhtang
> * On line 99, the citation to Tensor Toolbox is malformed - it seems there is a "and " missing before "others";

@melissawm My colleague @dcamps did some research and found that having an `and others` term at the end of the author list is a [BibTeX convention](https://tex.stackexchange.com/a/123607), and would be rendered as *et. al.* at compile time. However, [Whedon](https://whedon.theoj.org/) is apparently not supporting this convention. Is there any way that we can work around this?
