### 2022-03-17 9750ba7
update dependency to fix documentation tests
