### 2022-03-24 8bd71cd
Update bibtex entry for `TensorOperations.jl`.
