+++ paper.bib
@@ -7,11 +7,31 @@
 }
 
 @software{to,
-  author = {Jutho Haegeman},
-  title = {TensorOperations.jl},
-  url = {https://github.com/Jutho/TensorOperations.jl},
-  version = {3.2.3},
-  date = {2021-08-11}
+  author       = {Jutho and
+                  ho-oto and
+                  maartenvd and
+                  getzdan and
+                  Leo and
+                  Dilum Aluthge and
+                  Spencer Lyon and
+                  Alexander Morley and
+                  Austin Privett and
+                  Dmitri Iouchtchenko and
+                  Elliot Saba and
+                  Frank Otto and
+                  Jim Garrison and
+                  Julia TagBot and
+                  Katharine Hyatt and
+                  Marcus P S and
+                  Markus Hauru and
+                  Mason Protter},
+  title        = {Jutho/TensorOperations.jl: v3.2.3},
+  month        = aug,
+  year         = 2021,
+  publisher    = {Zenodo},
+  version      = {v3.2.3},
+  doi          = {10.5281/zenodo.5180716},
+  url          = {https://doi.org/10.5281/zenodo.5180716}
 }
 
 @software{einops,
