### 2022-03-24 #4198 @yhtang
So far we have addressed 3 of the 4 issues as below. For the remaining one, Could we get some technical support regarding the compilation system that Whedon uses?

> - [x] it seems you are using a slightly modified version of the [BSD-3 clause LBNL License](https://opensource.org/BSD-3-Clause-LBNL) - can you remove the added text at the beginning of  the first line so that the standard license text is used?
> - [ ] On line 99, the citation to Tensor Toolbox is malformed - it seems there is a "and " missing before "others";
> - [x] Can you clarify the citation for TensorOperations.jl? I see one author cited but several contributors in the github repository, and no citation instructions.
> - [x] Could you add the doi for TensorLy? https://dl.acm.org/doi/abs/10.5555/3322706.3322732

> @melissawm My colleague @campsd did some research and found that having an and others term at the end of the author list is a [BibTeX convention](https://tex.stackexchange.com/a/123607), and would be rendered as et. al. at compile time. However, [Whedon](https://whedon.theoj.org/) is apparently not supporting this convention. Is there any way that we can work around this?
