### 2022-03-25 2df5453
driver+plugin design of the gradient descent loop
