+++ funfact/gradient_descent/__init__.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+from ._base import GradientDescentState, GradientDescentPlugin
+from ._driver import gradient_descent
+from ._plugins import (
+    gradient_descent_plugin,
+    snapshot,
+    walltime
+)
