### 2022-03-25 0e2d5f4
migrate the `factorize` method to use the new gradient descent driver
