+++ funfact/__init__.py
@@ -14,6 +14,7 @@ from . import initializers
 from . import conditions
 from .lang._semiring import minplus, logsumexp, viterbi
 from . import parametrized
+from .gradient_descent import gradient_descent_plugin
 
 __all__ = [
     'use',
@@ -38,7 +39,8 @@ __all__ = [
     'minplus',
     'logsumexp',
     'viterbi',
-    'parametrized'
+    'parametrized',
+    'gradient_descent_plugin'
 ]
 
 
