+++ funfact/gradient_descent/_plugins.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 import time
-import dill
 from ._base import GradientDescentPlugin
 
 
