+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 1.0.3 (2022-03-27)
+
+# Bug fixes: #232.
+
 ## 1.0.2 (2022-03-27)
 
 # Bug fixes: #228.
