### 2022-03-28 bb84976
docstring for the `plugins` argument.
