### 2022-04-20 a562284
simple circuits and gates
