+++ funfact/quantum/_qgates.py
@@ -9,28 +9,31 @@ from funfact.parametrized import Generator
 from funfact.lang._tsrex import TsrEx, tensor
 from funfact.lang._ast import Primitives as P
 from funfact.lang._terminal import ParametrizedAbstractTensor
-from funfact.lang._special import eye
-from ._special import proj0, proj1
+from funfact.lang._special import eye, proj0, proj1
 from funfact.conditions import Unitary
-# TODO: how to handle dtype? Should always be complex!
+from funfact.util.integral import check_bounded_integral
 
 
 class Gate(ABC):
-    '''A common base class for all gate objects.'''
+    '''A common abstract base class for all gate objects.'''
     def __init__(self, qubits: List[Integral]):
-        '''A gate stores the minimum and maximum qubits it acts on.'''
-        for i, q in enumerate(qubits):
-            if not (isinstance(q, Integral) and q >= 0):
-                raise RuntimeError(
-                    f"Qubit indices must be non-negative integers, got {q} for"
-                    f" entry {i}."
-                )
-        qubits.sort()
-        self._qubits = qubits
+        '''A gate stores the minimum and maximum index of thequbits it acts
+        on.
+
+        Args:
+            qubits: List[Integral]:
+                The zero-based minimum and maximum indices of the qubits that
+                the gate acts on.
+        Returns:
+            self:
+                A Gate object.
+        '''
+        check_bounded_integral(qubits, minv=0)
+        self._qubits = tuple(sorted(qubits))
 
     @abstractmethod
-    def to_tensor(self):
-        '''A gate can be converted to a tensor, either parametrized or not.'''
+    def to_tsrex(self):
+        '''Tensor expression representation of the gate.'''
         pass
 
     @abstractmethod
@@ -40,38 +43,52 @@ class Gate(ABC):
 
     @property
     def qubits(self):
+        '''The minimum and maximum indices of the qubits that the gate acts
+        on.'''
         return self._qubits
 
 
 class OneQubitGate(Gate):
     '''Abstract base class for gates acting on exactly one qubit.'''
     def __init__(self, qubit: Integral):
-        if not (isinstance(qubit, Integral)):
-            raise RuntimeError(
-                f"Qubit index of one qubit gate must be integer, got {qubit}."
-            )
+        '''
+        Args:
+            qubit: integer
+                The zero-based index of the qubit the gate acts on.
+        '''
+        check_bounded_integral(qubit)
         super().__init__(list((qubit,)))
 
 
 class OneQubitUnitary(OneQubitGate):
     '''A dense one qubit unitary gate.'''
-    def __init__(self, qubit: int, initializer=None):
+    def __init__(self, qubit: Integral, label=None, initializer=None):
+        '''
+        Args:
+            qubit: integer
+               The zero-based index of the qubit the gate acts on.
+            label: string
+                Label of the quantum gate.
+            initializer
+                Initialization distribution
+        '''
         self._initializer = initializer
+        self._label = label
         super().__init__(qubit)
 
-    def to_tensor(self):
+    def to_tsrex(self):
         return tensor(
             self.label, 2, 2, initializer=self._initializer, prefer=Unitary()
         )
 
     @property
     def label(self):
-        return 'Q_1'
+        return self._label if self._label else 'Q_1'
 
 
 class PauliX(OneQubitGate):
     '''A one qubit Pauli-X gate.'''
-    def to_tensor(self):
+    def to_tsrex(self):
         return tensor(self.label, np.array([[0, 1], [1, 0]]))
 
     @property
@@ -81,7 +98,7 @@ class PauliX(OneQubitGate):
 
 class PauliY(OneQubitGate):
     '''A one qubit Pauli-Y gate.'''
-    def to_tensor(self):
+    def to_tsrex(self):
[…]
