### 2022-04-26 #4198 @melissawm
Sure - if they are not already a JOSS reviewer from our spreadsheet, you may also encourage them to [add their names there as potential reviewers for the future]([joss.theoj.org/reviewer-signup.html](https://joss.theoj.org/reviewer-signup.html)) 😄
