+++ funfact/quantum/_qgates.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-from abc import ABC, abstractmethod
+from abc import ABC
 from numbers import Integral
 from typing import List
 from funfact import active_backend as ab
@@ -11,35 +11,25 @@ from funfact.lang._ast import Primitives as P
 from funfact.lang._terminal import ParametrizedAbstractTensor
 from funfact.lang._special import eye, proj0, proj1
 from funfact.conditions import Unitary
-from funfact.util.integral import check_bounded_integral
 
 
-class Gate(ABC):
-    '''A common abstract base class for all gate objects.'''
-    def __init__(self, qubits: List[Integral]):
-        '''A gate stores the minimum and maximum index of thequbits it acts
-        on.
+_proj0 = proj0()
+_proj1 = proj1()
 
-        Args:
-            qubits: List[Integral]:
-                The zero-based minimum and maximum indices of the qubits that
-                the gate acts on.
-        Returns:
-            self:
-                A Gate object.
-        '''
-        check_bounded_integral(qubits, minv=0)
-        self._qubits = tuple(sorted(qubits))
 
-    @abstractmethod
+class _PlacedGate:
+    '''A placed gate wraps a gate and contains the qubit position
+    information.'''
+    def __init__(self, gate, qubits: List[Integral]):
+        self._gate = gate
+        self._qubits = tuple(qubits)
+
     def to_tsrex(self):
-        '''Tensor expression representation of the gate.'''
-        pass
+        return self._gate.to_tsrex(self._qubits)
 
-    @abstractmethod
+    @property
     def label(self):
-        '''A gate has a label.'''
-        pass
+        return self._gate.label
 
     @property
     def qubits(self):
@@ -47,97 +37,112 @@ class Gate(ABC):
         on.'''
         return self._qubits
 
-
-class OneQubitGate(Gate):
-    '''Abstract base class for gates acting on exactly one qubit.'''
-    def __init__(self, qubit: Integral):
-        '''
-        Args:
-            qubit: integer
-                The zero-based index of the qubit the gate acts on.
-        '''
-        check_bounded_integral(qubit)
-        super().__init__(list((qubit,)))
+    def __repr__(self) -> str:
+        print(self.label)
+        return fr'{self.label}_{{q:{self._qubits}}}'
 
 
-class OneQubitUnitary(OneQubitGate):
+class Gate(ABC):
+    '''A common abstract base class for all gate objects. Gates have a label
+    and a tensor expression representation.'''
+    def __init__(self):
+        pass
+
+    def to_tsrex(self, qubits=None):
+        '''Tensor expression representation of the gate.'''
+        return self._tsrex
+
+    @property
+    def label(self):
+        '''A gate has a label.'''
+        return self._label
+
+    def __matmul__(self, qubits: List[Integral]):
+        if isinstance(qubits, int):
+            qubits = list((qubits,))
+        return _PlacedGate(self, qubits)
+
+    def __repr__(self) -> str:
+        return self.label
+
+
+class OneQubitUnitary(Gate):
     '''A dense one qubit unitary gate.'''
-    def __init__(self, qubit: Integral, label=None, initializer=None):
+    def __init__(self, label=None, initializer=None):
         '''
         Args:
-            qubit: integer
-               The zero-based index of the qubit the gate acts on.
             label: string
                 Label of the quantum gate.
             initializer
                 Initialization distribution
         '''
         self._initializer = initializer
-        self._label = label
-        super().__init__(qubit)
-
-    def to_tsrex(self):
-        return tensor(
[…]
