### 2022-05-09 #4198 @yhtang
Dear @melissawm: I reached out to the primary contributor of the [torhcquantum](https://github.com/mit-han-lab/torchquantum) package, [hanrui-wang,](https://github.com/Hanrui-Wang), who is willing to review our submission. Could you please let us know your thoughts? Thanks!
