### 2022-06-08 #4198 @melissawm
Hello @fabian-sp, are willing to review this submission to JOSS, following our review guidelines: [joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html) ? Thank you!
