### 2022-07-04 #4502 @fabian-sp
I will open an issue in the repository for every topic/question that I encounter. Hope this is okay for everyone.
