### 2022-07-04 yhtang/FunFact#241 @fabian-sp
I think the reason for this is that I installed `FunFact` in an empty `conda` environment. So I have neither `torch` nor `jax` and then all autograd functions can of course not be used. Maybe this should be stated somewhere in the `README`? In the `requirements.txt` there is none of those packages specified.
