### 2022-07-04 yhtang/FunFact#242 @fabian-sp
JOSS review: examples and numerical validation
This is related to the review of `FunFact` for JOSS (see https://github.com/openjournals/joss-reviews/issues/4502)

I understand that your package can do (or approximate) arbitrary tensor decompositions with optimization. As the underlying properties of these optimization problems (convexity, uniqueness of solutions etc.) are not fully clear to me, I think it would be very nice to have some certainty that this approach works for standard decomposition such as SVD or the Tucker decomposition (i.e. that the optimizer converges to a reasonably good decomposition).

Moreover, for these special cases one could compare the result of `FunFact` with standard libraries for SVD or the libraries you mention for specialized tensor decomposition.

I see that this example (https://funfact.readthedocs.io/en/latest/examples/matrix-approximation/) already goes into this direction, but I think it might be nice to have a more numerical comparison of the decomposed factors. I haven't found any unit tests doing this, but maybe I just did not find them :)

Thanks a lot!
