### 2022-07-05 yhtang/FunFact#241 @fabian-sp
Update on this: after installing `torch` the unit tests are succesfull. I would still recommend to make it more prominent in the `README` (and/or `requirements.txt`) that without having `torch` or `jax` installed, a major part of the functionalities of the package cannot be used.
