### 2022-07-05 yhtang/FunFact#241 @yhtang
Thanks for the feedback! We mentioned the installation of FunFact with specific autograd backends in the [Installation](https://funfact.readthedocs.io/en/latest/pages/installation/#autograd-backends) page of our online docs. Perhaps there could be a better place for this information? What would be your suggestions?
