### 2022-07-05 yhtang/FunFact#242 @yhtang
Thanks for pointing this out. We can set up some self-consistency tests to check if FunFact can find solutions close enough to those found by 'standard solvers' for convex problems such as SVD and eigendecomposition.
