### 2022-07-06 yhtang/FunFact#241 @fabian-sp
Ah, thank you, indeed I did overlook this. Maybe put a link to this page on the Github `README` and the landing page of your documentation (that's where I usually look first if I would want to install your package)?
