### 2022-07-06 #4502 @fabian-sp
I have a general question regarding the license: on Github I see on the right-hand side only a "View license" where I typically would see the license type (e.g. BSD-3-Clause license). As I am not 100% firm on licencing legal background, I would like to ask @melissawm  what exactly needs to be fulfilled to check off the License point? Strictly speaking, Github seems to not recognize the license of this repo as a BSD-3-Clause license.
