### 2022-07-20 #4502 @yhtang
> Hello, @fabian-sp ! We had a [discussion about this in the pre-review](https://github.com/openjournals/joss-reviews/issues/4198#issuecomment-1070335568). @yhtang can you confirm that this is correct? Thank you!

Correct. Our LBNL variant of the BSD license is an [OSI-approved open-source license](https://opensource.org/licenses/alphabetical). We had a minor format issue during the pre-review, and that was already fixed.
