### 2022-08-12 6e3ae06
more installation instructions on README and landing page
