### 2022-08-12 4b8a63a
nvec -> vec_size
