### 2022-08-19 #4502 @fabian-sp
Hi @melissawm, I had opened three issues with questions or possible additional examples for the package. One of them has been adressed already. I would wait for an answer on the other two questions I had before checking all the remaining points in my checklist.
