### 2022-09-15 dd896e1
Correct the parameter name in README example.
Closes #245.
