### 2022-10-06 #4502 @yhtang
> Hi folks, any updates on this review? Thanks!

Hi @melissawm, we are currently working on a recently discovered bug, which is related to the self-consistency check that @fabian-sp raised. It turns out to have a deeper cause and hence would require a little longer to fix. I'll update back once we have that taken care of. Thanks for asking!
